requirements.txt
certifi==2018.4.16
chardet==3.0.4
docopt==0.6.2
idna==2.7
join==0.1.1
Markdown==2.6.11
numpy==1.14.5
Pillow==5.1.0
pipreqs==0.4.9
pyperclip==1.6.2
requests==2.19.1
urllib3==1.23
yarg==0.1.9

